# nexus5proxy/lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NexusProxy API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cart/add:
    post:
      operationId: addToCart
      tags: [cart]
      summary: Reserve a proxy in the cart
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AddToCartBody"
      responses:
        "201":
          description: Reserved
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/CartItem"
        "409":
          description: Proxy unavailable
        "410":
          description: Proxy gone
  /cart/{id}:
    delete:
      operationId: removeFromCart
      tags: [cart]
      summary: Release a cart reservation
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Removed
...
components:
  schemas:
...
    CartItem:
      type: object
      properties:
        id: { type: string }
        proxyId: { type: string }
        expiresAt: { type: string }
        createdAt: { type: ["string", "null"] }
        ip: { type: ["string", "null"] }
        port: { type: ["integer", "null"] }
        proxyType: { type: ["string", "null"] }
        country: { type: ["string", "null"] }
        city: { type: ["string", "null"] }
        priceCents: { type: ["integer", "null"] }
      required: [id, proxyId, expiresAt]
    AddToCartBody:
      type: object
      properties:
        proxyId: { type: string }
      required: [proxyId]
...